==> notebooks/01-01_download_data.ipynb <==
## Step 1: Downloading all data
Downloading from eggNOG and gold should take around 30min-1h given a fast internet connection, in total there are about 150GB of data.

e6_url = "http://eggnog6.embl.de/download/eggnog_6.0/"
all_trees_path = e6_url + "e6.all_raw_trees_and_algs.tsv"
og2seqs_species_path = e6_url + "e6.og2seqs_and_species.tsv"
seq2ogs_path = e6_url + "e6.seq2ogs.tsv"
gold_url = "https://gold.jgi.doe.gov/download?mode=site_excel"


import os
parent_dir = os.path.dirname(os.getcwd())
data_dir = os.path.join(parent_dir, 'data')

Use bash command wget to download all files. For GOLD we need to extract a single sheet from the file, for that xlsx2csv needs to be installed via pip.

%%bash -s "$data_dir" "$all_trees_path" "$og2seqs_species_path" "$seq2ogs_path" "$gold_url"
pip install xlsx2csv
mkdir -p $1

mkdir -p $1/eggnog6
wget $2 -P $1/eggnog6
wget $3 -P $1/eggnog6
wget $4 -P $1/eggnog6

mkdir -p $1/gold
wget $5 -O $1/gold/goldData.xlsx

%%bash -s "$data_dir"
xlsx2csv $1/gold/goldData.xlsx -s 4 > $1/gold/goldData_organisms.csv
# rm $1/goldData.xlsx

All e6 files are quite large and take long to work with, so the first step is to extract a subset containing only Bacteria data:

%%bash -s "$data_dir"
# Select bacterial OGs, and bacterial trees.
# The second column in the trees file should be "2". There are no other single-character columns.
# and then the first and third columns can be used to prepare a subset COG->tree mapping file.
rg "\t2\t" $1//eggnog6/e6.all_raw_trees_and_algs.tsv > $1//eggnog6/e6.all_raw_trees_and_algs_bacteria.tsv
# similarly here the first column is the taxonomic grouping
rg -P '^2\t' $1//eggnog6/e6.og2seqs_and_species.tsv > $1//eggnog6/e6.og2seqs_and_species_bacteria.tsv
