==> notebooks/03-compile_outputs.ipynb <==
# Compile Results

All the program outputs are in different formats. The code here compiles them into a common format for easier comparison.

# to suppress warning from ete3 because it's not up to date with python3.12
import warnings
warnings.filterwarnings("ignore", category=SyntaxWarning)

import ete3
import pandas as pd
import os
import sys

# in ../code we have compile_results.py which contains all the functions we need
sys.path.append(os.path.join(os.path.dirname(os.getcwd()), 'code'))
import compile_results

genome_tree_labeled_path = "../data/genome_tree/genome_tree.iqtree.treefile.rooted.labeled"
genome_tree_labeled = ete3.Tree(genome_tree_labeled_path, format=1)

## Count

# compile the results for the count changes in the gene dynamics 
print("\033[1m" + "Compiling count changes in gene dynamics" + "\033[0m")
count_output_dir = "../data/inferences/gene_dynamics/Count/"
compile_results.compile_count_changes(count_dir=count_output_dir,
                                    input_tree_filepath=genome_tree_labeled_path,
                                    compiled_results_dir="../data/compiled_results/",
                                    var_str='gene', var_name_str='nog_id')
print("-----------------------------------")
# compile the results for the count changes in the ecosystem type dynamics
print("\033[1m" + "Compiling count changes in ecosystem type dynamics" + "\033[0m")
count_output_dir = "../data/inferences/ecotype_dynamics/Count/"
compile_results.compile_count_changes(count_dir=count_output_dir,
                                    input_tree_filepath=genome_tree_labeled_path,
                                    compiled_results_dir="../data/compiled_results/",
                                    var_str='ecotype', var_name_str='ecotype')
print("-----------------------------------")
# compile the results for the count changes in the ecosystem subtype dynamics
print("\033[1m" + "Compiling count changes in ecosystem subtype dynamics" + "\033[0m")
count_output_dir = "../data/inferences/ecosubtype_dynamics/Count/"
compile_results.compile_count_changes(count_dir=count_output_dir,
                                    input_tree_filepath=genome_tree_labeled_path,
                                    compiled_results_dir="../data/compiled_results/",
                                    var_str='ecosubtype', var_name_str='ecosubtype')
print("-----------------------------------")
print("All results compiled and saved in ../data/compiled_results/")
